refactor(gateway): isolate companion state contract
