+++ src/gateway/session-companion-state.ts
@@ -1,0 +1,38 @@
+import type { SessionCompanionExchange } from "../../packages/gateway-protocol/src/schema/sessions.js";
+
+export type SessionCompanionSeedMessage = {
+  role: "user" | "assistant";
+  text: string;
+  ts: number;
+};
+
+export type SessionCompanionThread = {
+  exchanges: SessionCompanionExchange[];
+  seed: {
+    messages: SessionCompanionSeedMessage[];
+    digestJson: string;
+  };
+  lastNoteSequence: number;
+  busy: boolean;
+  lastUsedAt: number;
+};
+
+export const SESSION_COMPANION_MAX_EXCHANGES = 24;
+export const SESSION_COMPANION_MAX_EXCHANGE_BYTES = 48 * 1024;
+
+function exchangeBytes(exchange: SessionCompanionExchange): number {
+  return Buffer.byteLength(exchange.question, "utf8") + Buffer.byteLength(exchange.answer, "utf8");
+}
+
+export function trimSessionCompanionExchanges(exchanges: SessionCompanionExchange[]): void {
+  let bytes = exchanges.reduce((total, exchange) => total + exchangeBytes(exchange), 0);
+  // Dropping the oldest exchange intentionally breaks the replay byte prefix;
+  // the count and byte caps take priority once a long-lived thread is bounded.
+  while (
+    exchanges.length > SESSION_COMPANION_MAX_EXCHANGES ||
+    bytes > SESSION_COMPANION_MAX_EXCHANGE_BYTES
+  ) {
+    const removed = exchanges.shift();
+    bytes -= removed ? exchangeBytes(removed) : 0;
+  }
+}
